lejog2021/requirements.txt
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dataclasses==0.6
flake8==3.8.4
Flask==1.1.2
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-datastore==2.1.0
googleapis-common-protos==1.52.0
gpx-cmd-tools==0.0.1
gpxpy==1.4.2
grpcio==1.34.0
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.16
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
